app_python/requirements.txt
fastapi
ngrok
openai
pydantic
pymongo
python-dotenv
uvicorn
langchain-openai
langchain_core
langchain_community
langserve
pandas
scikit-learn